Move JaxbContextProvider to business-impl module.
I am aware that business-impl is sort of an unexpected location
for this class to reside. Unfortunately it's the closest common
dependency of all the modules which need this class.

This should fix #26.
