+++ compass-database-initialization/pom.xml
@@ -45,11 +45,6 @@
 			<version>${version.jboss-logging}</version>
 		</dependency>
 		<dependency>
-			<groupId>org.eclipse.persistence</groupId>
-			<artifactId>org.eclipse.persistence.moxy</artifactId>
-			<version>${version.eclipselink}</version>
-		</dependency>
-		<dependency>
 			<groupId>javax</groupId>
 			<artifactId>javaee-api</artifactId>
 			<!-- version/scope is configured in root pom -->
